requirements.txt
# Binance API
ccxt>=4.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Technical indicators
pandas-ta>=0.3.14b

# Machine Learning
lightgbm>=4.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# NLP / Sentiment
transformers>=4.30.0
torch>=2.0.0

# Database (TimescaleDB/PostgreSQL)
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Dashboard
streamlit>=1.28.0
plotly>=5.15.0

# Environment & Config
python-dotenv>=1.0.0

# Async / WebSocket
aiohttp>=3.9.0
websockets>=12.0

# HTTP
requests>=2.31.0

# Logging
loguru>=0.7.0

# Scheduling
schedule>=1.2.0
